--- a/utils/datasets.py
+++ b/utils/datasets.py
@@ -27,7 +27,7 @@ help_url = 'https://github.com/ultralytics/yolov5/wiki/Train-Custom-Data'
 img_formats = ['bmp', 'jpg', 'jpeg', 'png', 'tif', 'tiff', 'dng']  # acceptable image suffixes
 vid_formats = ['mov', 'avi', 'mp4', 'mpg', 'mpeg', 'm4v', 'wmv', 'mkv']  # acceptable video suffixes
 logger = logging.getLogger(__name__)
-turbojpeg_flag = False
+turbojpeg_flag = True
 
 try:
     from turbojpeg import TurboJPEG