Dear Ultralytics:
    I noticed that Yolov5 suffer from performance drop when dealing with jpeg images. I currently update the dataset.py to support TurboJpeg packages to allow perf increase. Feel free to contact me if you have any questions. The code is tested under Ubuntu20 with pytorch1.6 on single card.