@glenn-jocher 2020-12-28
@cydiachen hi thank for the PR! Could you provide some profiling results before and after to measure speed improvements?