@cydiachen 2020-12-29
# Speed Benchmark Under Win64 platform

CPU: Ryzen7 4800H

RAM: 16GB

Storage: WD20SPZX - HDD

GPU: NVIDIA Quadro P620 with 4GB RAM

Default Configuration under Yolov5s with COCO128 dataset. 
Batch-size | Before Turbojpeg | After Turbojpeg
-- | -- | --
2 | 4.72it/s | 4.65it/s (-1.4%)
4 | 2.73it/s | 2.70it/s (-1.1%)
8 | 1.50it/s | 1.50it/s (-0.0%)
10 | 1.23it/s | 1.22it/s(-0.1%)

Default Configuration under yolov5s with PRIVATE dataset (4000*3000)

Batch-size | Before Turbojpeg | After Turbojpeg
-- | -- | --
2 | CUDA ERROR | CUDA ERROR
4 | 2.04it/s | 2.83it/s (+27.9%)
8 | 1.57it/s | 1.59it/s (+1.3%)
10 | 1.26it/s | 1.28it/s (+1.6%)

The performance boost introduced by TurboJPEG is highly correlated with the CPU, Batch_size of the training and the Original size of Input image.