@cydiachen 2020-12-29
> @cydiachen hi thank for the PR! Could you provide some profiling results before and after to measure speed improvements?

i WILL KEEP UPDATING PERFORMANCE INCREASE ON UBUNTU.