@cydiachen 2021-01-05
> @cydiachen thanks for the numbers. It seems for smaller images (like COCO) there is not much benefit to be had, though for larger images training at a smaller size (i.e. a dataset of 4k images trained at --img 640), there is a level of built-in inefficiency currently due to the resizing operation that is required at every image loading.
> 
> The recommended workaround for this is to `--cache` your data in RAM when training (images are cached at --img 640 in the above example, not at their native 4k size).
> `python train.py --cache`
> 
> Alternatively if you have insufficient RAM for the above task, you may want to simply resize your dataset to better approximate your training size. You could use this to do this:
> https://github.com/ultralytics/flickr_scraper/blob/master/utils/clean_images.py

Thx for your reply. I am currently suffering from a cold. I will be updating this perf benchmark with more details. The turbojpeg is widely used by OpenMMLAB's mmdetection, which may contributes to the performance boost at train time. I will try your advice.